+++ lib/shadcn_ui.dart
@@ -3,6 +3,15 @@
 import 'package:flutter_animate/flutter_animate.dart' as _animate show Effect;
 import 'package:intl/intl.dart' as intl show TextDirection;
 
+export 'package:boxy/boxy.dart';
+export 'package:boxy/flex.dart';
+export 'package:boxy/inflating_element.dart';
+export 'package:boxy/padding.dart';
+export 'package:boxy/redirect_pointer.dart';
+export 'package:boxy/render_boxy.dart';
+export 'package:boxy/scale.dart';
+export 'package:boxy/slivers.dart';
+export 'package:boxy/utils.dart';
 // Re-export collision libraries names
 export 'package:flutter_animate/flutter_animate.dart' hide Effect;
 // External libraries